requirements.txt
Django>=1.10.2
psycopg2>=2.6.2
djangorestframework>=3.5.3
dj-database-url
gunicorn
coreapi
